+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Updated the MSRV to 1.86 and edition 2024
+  [#666](github.com/edgehog-device-manager/edgehog-device-runtime/pull/666)
+
 ## [0.10.3] - 2025-12-19
 
 ### Fixed